backend/agents/requirements.txt
langchain-groq
langchain-core
langchain
langchain_community
pydantic
python-dotenv